allow cpu freq to be fractional
git-svn-id: https://svn.code.sf.net/p/pinmame/code/trunk@4961 6f51dfdd-75df-4273-8f3b-818b1e6aa8bf

+++ src/cpuexec.h
@@ -43,7 +43,7 @@ struct MachineCPU
 {
 	int			cpu_type;					/* index for the CPU type */
 	int			cpu_flags;					/* flags; see #defines below */
-	int			cpu_clock;					/* in Hertz */
+	double		cpu_clock;					/* in Hertz */
 	const void *memory_read;				/* struct Memory_ReadAddress */
 	const void *memory_write;				/* struct Memory_WriteAddress */
 	const void *port_read;
